+++ Sources/Internal/Extensions/String.swift
@@ -44,6 +44,15 @@ extension String {
         }
         return removingPrefix(prefix).addingPrefix(replacement)
     }
+
+    /// Returns a copy of the string after adding the given `suffix` if it's not already there.
+    public func addingSuffix(_ suffix: String) -> String {
+        if hasSuffix(suffix) {
+            return self
+        } else {
+            return self + suffix
+        }
+    }
     
     /// Returns a copy of the string after removing the given `suffix`, when present.
     public func removingSuffix(_ suffix: String) -> String {