+++ TestApp/Sources/Reader/CBZ/CBZModule.swift
@@ -28,8 +28,8 @@ final class CBZModule: ReaderFormatModule {
     }
     
     @MainActor
-    func makeReaderViewController(for publication: Publication, locator: Locator?, bookId: Book.Id, books: BookRepository, bookmarks: BookmarkRepository, highlights: HighlightRepository, resourcesServer: ResourcesServer) async throws -> UIViewController {
-        let cbzVC = CBZViewController(publication: publication, locator: locator, bookId: bookId, books: books, bookmarks: bookmarks)
+    func makeReaderViewController(for publication: Publication, locator: Locator?, bookId: Book.Id, books: BookRepository, bookmarks: BookmarkRepository, highlights: HighlightRepository) async throws -> UIViewController {
+        let cbzVC = try CBZViewController(publication: publication, locator: locator, bookId: bookId, books: books, bookmarks: bookmarks)
         cbzVC.moduleDelegate = self.delegate
         return cbzVC
     }