+++ TestApp/Sources/Reader/PDF/PDFModule.swift
@@ -31,9 +31,9 @@ final class PDFModule: ReaderFormatModule {
     }
     
     @MainActor
-    func makeReaderViewController(for publication: Publication, locator: Locator?, bookId: Book.Id, books: BookRepository, bookmarks: BookmarkRepository, highlights: HighlightRepository, resourcesServer: ResourcesServer) async throws -> UIViewController {
+    func makeReaderViewController(for publication: Publication, locator: Locator?, bookId: Book.Id, books: BookRepository, bookmarks: BookmarkRepository, highlights: HighlightRepository) async throws -> UIViewController {
         let preferencesStore = makePreferencesStore(books: books)
-        let viewController = PDFViewController(
+        let viewController = try PDFViewController(
             publication: publication,
             locator: locator,
             bookId: bookId,